Configure host prefix injection
This change allows for host prefix injection to be disabled through an
environment variable `AWS_DISABLE_HOST_PREFIX_INJECTION` or a shared
configuration file parameter `disable_host_prefix_injection`. This is
in addition to the existing Config parameter `inject_host_prefix`.

This change also changes the default value of the Config object's
`inject_host_prefix` from `True` to `None`. A property is used to
preserve the existing behavior when a user does not specify the
parameter to set it to `True`.

This preserves the behavior of a user setting inject_host_prefix=None
while keeping the logic internal.